lab7

+++ app/Http/Controllers/ArticleController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Article;
+use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
@@ -21,6 +23,7 @@ class ArticleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create', Article::class);
         return view('article.create');
     }
 
@@ -29,6 +32,7 @@ class ArticleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Article::class);
         $request->validate([
             'date' => 'required|date',
             'title' => 'required|min:10',
@@ -48,7 +52,8 @@ class ArticleController extends Controller
      */
     public function show(Article $article)
     {
-        return view('article.show', ['article'=>$article]);
+        $comments = Comment::where('article_id', $article->id)->get();
+        return view('article.show', ['article'=>$article, 'comments'=>$comments]);
     }
 
     /**
@@ -56,6 +61,7 @@ class ArticleController extends Controller
      */
     public function edit(Article $article)
     {
+        Gate::authorize('create', Article::class);
         return view('article.edit', ['article'=>$article]);
     }
 
@@ -82,6 +88,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        Gate::authorize('create', Article::class);
         $article->delete();
         return redirect()->route('article.index')->with('message','Delete successful');
     }
